♻️ Refactor email service
