+++ src/services/email/index.js
@@ -1,0 +1,2 @@
+export * from './news';
+export * from './passwordReset';
